refactor: Migrate `z_image_module_v3` from experimental to stable API
Updates imports to use stable `max.functional`, `max.tensor`, and `max.nn` modules instead of their experimental counterparts.

Renames `PixelGenerationContext` to `PixelContext` and moves `KVCacheStrategy` to legacy namespace.

Replaces `Qwen3Config` references with `Qwen3EncoderConfig` throughout the architecture.

Removes unused code including the `prepare_latents` method, denoising loop implementation details, and imports for `AutoencoderKL` and `FlowMatchEulerDiscreteScheduler`.

Updates attention and normalization imports to use legacy namespaces where appropriate.
