+++ max/python/max/pipelines/architectures/z_image_module_v3/model_config.py
@@ -20,8 +20,7 @@ from types import SimpleNamespace
 from max.dtype import DType
 from max.graph.weights import WeightData
 from max.nn import ReturnLogits
-from max.nn.kv_cache import KVCacheParams
-from max.pipelines.architectures.qwen3.model_config import Qwen3Config
+from max.nn.legacy.kv_cache import KVCacheParams
 from max.pipelines.lib import KVCacheConfig, MAXModelConfigBase, PipelineConfig
 from transformers.models.auto.configuration_auto import AutoConfig
 
@@ -311,7 +310,7 @@ class ZImageConfig(MAXModelConfigBase):
     vae_config: VAEConfig
     """VAE configuration."""
 
-    text_encoder_config: Qwen3Config
+    text_encoder_config: Qwen3EncoderConfig
     """Text encoder configuration."""
 
     transformer_config: TransformerConfig
@@ -330,11 +329,11 @@ class ZImageConfig(MAXModelConfigBase):
         kv_cache_config: KVCacheConfig,
         cache_dtype: DType,
     ) -> KVCacheParams:
-        # Delegate to Qwen3Config for language model parameters.
+        # Delegate to Qwen3EncoderConfig for language model parameters.
         llm_config = getattr(
             huggingface_config, "text_config", huggingface_config
         )
-        return Qwen3Config.get_kv_params(
+        return Qwen3EncoderConfig.get_kv_params(
             huggingface_config=llm_config,
             n_devices=n_devices,
             kv_cache_config=kv_cache_config,
@@ -343,22 +342,22 @@ class ZImageConfig(MAXModelConfigBase):
 
     @staticmethod
     def get_num_layers(huggingface_config: AutoConfig) -> int:
-        # Delegate to Qwen3Config for language model parameters.
+        # Delegate to Qwen3EncoderConfig for language model parameters.
         llm_config = getattr(
             huggingface_config, "text_config", huggingface_config
         )
-        return Qwen3Config.get_num_layers(llm_config)
+        return Qwen3EncoderConfig.get_num_layers(llm_config)
 
     @staticmethod
     def calculate_max_seq_len(
         pipeline_config: PipelineConfig, huggingface_config: AutoConfig
     ) -> int:
         """Calculate maximum sequence length for ZImage."""
-        # Delegate to Qwen3Config for language model parameters.
+        # Delegate to Qwen3EncoderConfig for language model parameters.
         llm_config = getattr(
             huggingface_config, "text_config", huggingface_config
         )
-        return Qwen3Config.calculate_max_seq_len(
+        return Qwen3EncoderConfig.calculate_max_seq_len(
             pipeline_config=pipeline_config,
             huggingface_config=llm_config,
         )
@@ -410,7 +409,7 @@ class ZImageConfig(MAXModelConfigBase):
             pipeline_config,
         )
 
-        # Create Qwen3Config for the text encoder
+        # Create Qwen3EncoderConfig for the text encoder
         # Use ReturnHiddenStates.SECOND_TO_LAST to get hidden_states[-2]
         # (second-to-last layer) for Z-Image conditioning - matching diffusers behavior
         # Disable prefix caching for diffusion model text encoders
@@ -424,7 +423,7 @@ class ZImageConfig(MAXModelConfigBase):
             enable_prefix_caching=False,
         )
 
-        text_encoder_config = Qwen3Config.generate(
+        text_encoder_config = Qwen3EncoderConfig.generate(
             pipeline_config,
             text_encoder_config,
             text_encoder_state_dict["llm_state_dict"],
